+++ public/css/style.css
@@ -65,21 +65,6 @@ ul {
   max-width: 430px;
   height: 60px;
   background-color: #71A304;
-<<<<<<< HEAD
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  border-top: 1px solid #ccc;
-  box-shadow: 0 -2px 5px rgba(0, 0, 0, 0.1);
-  z-index: 1000;
-}
-
-.bottom-nav ul {
-  display: flex;
-  justify-content: space-around;
-  padding: 10px 0;
-  margin: 0;
-=======
   display: flex;
   justify-content: space-around;
   align-items: center;
@@ -93,7 +78,6 @@ ul {
   padding: 0;
   display: flex;
   justify-content: space-around;
->>>>>>> feature/growmarket-seoyeon
 }
 
 .bottom-nav a {
@@ -105,18 +89,8 @@ ul {
   font-weight: 500;
   font-size: 12px;
   color: #333;
-<<<<<<< HEAD
-  cursor: pointer;
-}
-
-.bottom-nav a i {
-  font-size: 18px;
-  margin-bottom: 2px;
-  color: #333;
-=======
 }
 .bottom-nav a i {
   font-size: 18px;
   margin-bottom: 2px;
->>>>>>> feature/growmarket-seoyeon
 }
